add setting of services name into sso.xml - rename dir - rename text key pem

+++ login.sh
@@ -1,8 +1,28 @@
 #!/bin/bash
 
-readonly SSO_TEST_XML="./resources/sso.test.xml"
+readonly AUTH_SERVER_CN="www.example.com"
 
-readonly EXAMPLE_KEY_PEM="./resources/www.example.com.key.pem"
+readonly SERVICE_CN="fe-webapp-bff"
+
+readonly SSO_TEST_XML="./test/sso.test.xml"
+
+readonly EXAMPLE_PRIVATE_KEY_PEM="./test/${AUTH_SERVER_CN}.privatekey.pem"
+
+readonly EXAMPLE_PUBLIC_KEY_PEM="./crypto-config/${AUTH_SERVER_CN}.publickey.pem"
+
+if [[ ! -f $EXAMPLE_PRIVATE_KEY_PEM ]]; then
+   echo "ERROR: file [$EXAMPLE_PRIVATE_KEY_PEM] not found !!!"
+   exit 1
+fi
+
+if [[ ! -f $EXAMPLE_PUBLIC_KEY_PEM ]]; then
+    openssl rsa -in "$EXAMPLE_PRIVATE_KEY_PEM" -pubout -out "$EXAMPLE_PUBLIC_KEY_PEM"
+fi
+
+if [[ ! -f $EXAMPLE_PUBLIC_KEY_PEM ]]; then
+   echo "ERROR: file [$EXAMPLE_PUBLIC_KEY_PEM] not found !!!"
+   exit 1
+fi
 
 rm -f ./tmp/sso.xml
 rm -f ./tmp/sso.xml.base64
@@ -16,9 +36,12 @@ exp_time=$(( gen_time + 60 ))
 readonly exp_time
 
 sed "s/{{GEN_TIME}}/${gen_time}/" "$SSO_TEST_XML" | \
-sed "s/{{EXP_TIME}}/${exp_time}/" > ./tmp/sso.xml
+sed "s/{{EXP_TIME}}/${exp_time}/" | \
+sed "s/{{AUTH_SERVER_CN}}/${AUTH_SERVER_CN}/" | \
+sed "s/{{SERVICE_CN}}/${SERVICE_CN}/g" \
+> ./tmp/sso.xml
 
-openssl dgst -sha256 -sign "$EXAMPLE_KEY_PEM" -out ./tmp/sign ./tmp/sso.xml
+openssl dgst -sha256 -sign "$EXAMPLE_PRIVATE_KEY_PEM" -out ./tmp/sign ./tmp/sso.xml
 
 for file in sso.xml sign; do
     base64 -w 0 "./tmp/$file" | sed 's/+/-/g' | tr -d '=' > "./tmp/$file.base64"
